10 feat implement login form with validation (#57)
* refactor: Update background image URL in layout component

* refactor: Update background image URL in layout component

* feat: Add bcryptjs and jsonwebtoken dependencies

This commit adds the `bcryptjs` and `jsonwebtoken` dependencies to the project. These packages are needed for implementing secure authentication and authorization functionality.

* refactor: Add react-hot-toast for displaying notifications

This commit adds the `react-hot-toast` package to the project and integrates it into the layout component. The `Toaster` component from `react-hot-toast` is added to display notifications at the bottom-right corner of the page. This will improve the user experience by providing visual feedback for important actions or events.

* feat: Add login schema for validation

* feat: Add LoginForm component for login page

This commit adds a new component called LoginForm to the login page. The LoginForm component includes input fields for email and password, as well as a submit button. It also handles form validation and submission using the react-dom useFormStatus hook. This addition improves the structure and functionality of the login page, making it easier for users to login to the application.

* feat: Add PrismaClient configuration for database connection

This commit adds the PrismaClient configuration in the `src/lib/prisma.ts` file. The `PrismaClient` is imported from the `@prisma/client` package. The configuration checks the `NODE_ENV` environment variable and initializes the `prisma` instance accordingly. In production, a new `PrismaClient` instance is created, while in development, the `prisma` instance is stored globally to avoid multiple instantiations. This addition enables the application to connect to the database using Prisma and perform database operations.

* refactor: Remove unused React import in landing_footer component

* refactor: Remove unused React import in login-action.ts

* feat: Update npm dependencies for react-hook-form and @hookform/resolvers

* refactor: Update email validation error message in LoginSchema

* refactor: Update CustomButton component styles

This commit updates the styles of the CustomButton component in the src/components/ui/custom_button.tsx file. The buttonStyle, iconStyle, and color variables are modified to improve the visual appearance of the button. This refactor enhances the overall user experience by providing a more visually appealing button component.